Add DashboardContext
Added Dashboard Context and passed dashboard type prop into it, so that lower level components can access it. Fixed links in responsive sidebar.
