+++ src/components/context/DashboardContext.tsx
@@ -1,0 +1,14 @@
+import {createContext, useContext} from "react";
+import {UserType} from "../layouts/DashboardLayout";
+
+interface DashboardContextProps {
+  type: UserType,
+}
+
+const DashboardContext = createContext<DashboardContextProps>({
+  type: "doctor",
+});
+
+export const useDashboardContext = () => useContext(DashboardContext);
+
+export default DashboardContext;
